Split SignTxView into SendTransferApproval and SignPsbtApproval (#47)
